requirements.txt
Django==4.2.4
djangorestframework==3.14.0
pre-commit==3.3.3
psycopg2-binary==2.9.7
gunicorn==21.2.0
PyJWT==2.8.0
drf-spectacular==0.26.4
